fixed imports/tests
